Move Gradle API related codegen code to :build-logic:kotlin-dsl
